Cleanup `size_t`: Fewer `#pragma warning` directives
 * Remove various `#pragma warning` directives that relate to
	signed/unsigned

	4018 - signed/unsigned mismatch
	4100 - unreferenced formal parameter
	4189 - local variable initialized but unreferenced
	4244 - argument conversion, loss of data
	4355 - `this` used in base member list
	4512 - assignment operator could not be generated
	4520 - non-existing error !!
	4711 - inlining notice (off by default)

 * use `size_t` as appropriate to eliminate warnings exposed above.
 * use `static_cast<int>` as appropriate

Additional, unrelated ...
 * `_AFX_NO_MFC_CONTROLS_IN_DIALOGS`was used twice for 'X64 Release'
	configuration.

+++ Externals/poco/Foundation/src/PurgeStrategy.cpp
@@ -128,7 +128,7 @@ void PurgeByCountStrategy::purge(const std::string& path)
 {
 	std::vector<File> files;
 	list(path, files);
-	while (files.size() > _count)
+	while (static_cast<int>(files.size()) > _count)
 	{
 		std::vector<File>::iterator it = files.begin();
 		std::vector<File>::iterator purgeIt = it;
