+++ Src/PluginsListDlg.cpp
@@ -104,7 +104,7 @@ void PluginsListDlg::AddPluginsToList(const wchar_t *pluginEvent, const String& 
 	PluginArray * piPluginArray = 
 		CAllThreadsScripts::GetActiveSet()->GetAvailableScripts(pluginEvent);
 
-	for (int iPlugin = 0 ; iPlugin < piPluginArray->size() ; iPlugin++)
+	for (size_t iPlugin = 0 ; iPlugin < piPluginArray->size() ; iPlugin++)
 	{
 		const PluginInfoPtr& plugin = piPluginArray->at(iPlugin);
 		int ind = m_list.InsertItem(m_list.GetItemCount(), plugin->m_name.c_str());
